client/requirements.txt
soundfile
simpleaudio
piper
grpcio
grpcio-tools
protobuf
python-dotenv
